Move prodcut tour code to theme's override
